Completely rewrite the config system.
Replace ClothConfig / AutoConfig with a YetAnotherConfigLib based
config approach. This allows to do some cool config screen magic:

We now render a preview of the halo in the config screen. This preview
updates as the config is saved. Each segment of the preview can be
clicked to open the options for that component.